[LRN] Limit textBlock.fuseChildren to text nodes
Prevents it from trying to fuse the cursor span, causing an error.

This is an issue with mathquill dev - submitted an issue and a pull
request at https://github.com/mathquill/mathquill/issues/429

When/if this PR gets merged into dev, we should remove this commit and
rebase on top of dev.

LRN-4213